+++ libs/vre/shared/app-ui/src/index.ts
@@ -3,3 +3,8 @@ export * from './lib/dialog/dialog.service';
 export * from './lib/dialog/dialog-sizes.constant';
 export * from './lib/dialog/confirm-dialog.component';
 export * from './lib/pager/pager.component';
+export * from './lib/time-input/seconds-to-time-string';
+export * from './lib/time-input/time-format.directive';
+export * from './lib/time-input/time-input.component';
+export * from './lib/time-input/time-string-to-seconds';
+export * from './lib/time-input/time-validator';
