+++ libs/vre/shared/app-ui/src/lib/time-input/seconds-to-time-string.ts
