+++ libs/vre/shared/app-ui/src/lib/time-input/time-string-to-seconds.ts
