+++ libs/vre/shared/app-ui/src/lib/time-input/time-validator.ts
