+++ libs/vre/shared/app-ui/src/ui.components.ts
@@ -1,5 +1,7 @@
 import { ConfirmDialogComponent } from './lib/dialog/confirm-dialog.component';
 import { PagerComponent } from './lib/pager/pager.component';
+import { TimeFormatDirective } from './lib/time-input/time-format.directive';
+import { TimeInputComponent } from './lib/time-input/time-input.component';
 
-export const UiComponents = [ConfirmDialogComponent];
+export const UiComponents = [ConfirmDialogComponent, TimeInputComponent, TimeFormatDirective];
 export const UiStandaloneComponents = [PagerComponent];
